Programs to rename a batch of photos and copy out the good ones.
RenameAll renames every photo in the current directory.  It takes photos
in alphabetical order and removes gaps in the numbering.

Copy takes in a file called 'good ones.txt', which should just have
numbers of the photos that are good, and creates a copy of each photo
and puts them in a new folder.