Renaming happens twice, first time uses a random prefix.
Renaming files can overwrite existing files if you try to rename a file
to a filename that already exists. Solution is to first create a random
prefix and check that no files start with that prefix. Rename each photo
to that prefix and then rename them to the new prefix and number.